+++ compilers/8c/swt.c
@@ -344,7 +344,7 @@ void
 zname(Biobuf *b, Sym *s, int t)
 {
     char *n;
-    ulong sig;
+    uint32 sig;
 
     if(debug['T'] && t == D_EXTERN && s->sig != SIGDONE && s->type != types[TENUM] && s != symrathole){
         sig = sign(s);
@@ -375,7 +375,7 @@ zname(Biobuf *b, Sym *s, int t)
 void
 zaddr(Biobuf *b, Adr *a, int s)
 {
-    long l;
+    int32 l;
     int i, t;
     char *n;
     Ieee e;
@@ -443,10 +443,10 @@ zaddr(Biobuf *b, Adr *a, int s)
 /*e: function [[zaddr]](x86) */
 
 /*s: function [[align]](x86) */
-long
-align(long i, Type *t, int op)
+int32
+align(int32 i, Type *t, int op)
 {
-    long o;
+    int32 o;
     Type *v;
     int w;
 
@@ -511,8 +511,8 @@ align(long i, Type *t, int op)
 /*e: function [[align]](x86) */
 
 /*s: function [[maxround]] */
-long
-maxround(long max, long v)
+int32
+maxround(int32 max, int32 v)
 {
     v = round(v, SZ_LONG);
     if(v > max)
